refactoring: solution sorting test moved to fw package
